requirements.txt
Django==1.9
argparse==1.2.1
django-taggit==0.17.6
wsgiref==0.1.2
